backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.0
faiss-cpu==1.7.4
numpy==1.26.2
pydantic==2.5.1
python-multipart==0.0.6
googletrans-py==4.0.0
sentence-transformers==2.2.2
chromadb==0.4.22